requirements.txt
Pillow==3.0.0
PyYAML==3.11
Twisted==15.4.0
autobahn==0.10.9
click==5.1
httplib2==0.9.2
numpy==1.10.1
pytz==2015.7
requests==2.8.1
selenium==2.48.0
six==1.10.0
txaio==2.0.4
txredisapi==1.3
ujson==1.33
wsgiref==0.1.2
zope.interface==4.1.3
